Moving cursors in the base class: DisplayPlot
Moved the cursors from oscilloscope to DisplayPlot and adapted the functionality.
